Renamed run-> setup, moved setup to scripts
